+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New in v11.7
 --------------------------
+- Add IFRA Categories explanation in usage page
 - This release may include more changes, for full details please refer to the CHANGELOG